add manager ticket, assign system, staff comment system
