+++ communitymanagement/src/main/java/communitymanagement/dao/StaffDao.java
@@ -1,19 +1,22 @@
 package communitymanagement.dao;
 
-import communitymanagement.model.Authorities;
-import communitymanagement.model.Staff;
-import communitymanagement.model.StaffCategory;
-import communitymanagement.model.User;
-import communitymanagement.model.UserType;
+import java.util.ArrayList;
+import java.util.List;
+
+import javax.persistence.criteria.CriteriaBuilder;
+import javax.persistence.criteria.CriteriaQuery;
+import javax.persistence.criteria.Root;
+
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
-import javax.persistence.criteria.CriteriaBuilder;
-import javax.persistence.criteria.CriteriaQuery;
-import javax.persistence.criteria.Root;
-import java.util.List;
+import communitymanagement.model.Authorities;
+import communitymanagement.model.Staff;
+import communitymanagement.model.StaffCategory;
+import communitymanagement.model.User;
+import communitymanagement.model.UserType;
 
 @Repository
 public class StaffDao {
@@ -67,6 +70,7 @@ public class StaffDao {
 
         return null;
     }
+ 
 
     public List<Staff> getStaffByCategory(String category) {
         StaffCategory staffCategory = null;
@@ -84,6 +88,31 @@ public class StaffDao {
 
         if (staffCategory != null) {
             return staffCategory.getStaff();
+        }
+
+        return null;
+    }
+    
+    public List<Staff> getStaffsByCategoryId(int staffCategoryId) {
+        StaffCategory staffCategory = null;
+        List<Staff> staffs = new ArrayList<>();
+        try (Session session = sessionFactory.openSession()) {
+            CriteriaBuilder criteriaBuilder = session.getCriteriaBuilder();
+            CriteriaQuery<StaffCategory> criteriaQuery = criteriaBuilder.createQuery(StaffCategory.class);
+            Root<StaffCategory> root = criteriaQuery.from(StaffCategory.class);
+            criteriaQuery.select(root).where(criteriaBuilder.equal(root.get("id"), staffCategoryId));
+            staffCategory = session.createQuery(criteriaQuery).getSingleResult();
+  
+            if (staffCategory != null) {
+            	staffs = staffCategory.getStaff();
+            }
+            session.getTransaction().commit();
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+
+        if (staffs != null) {
+            return staffs;
         }
 
         return null;
